moving OperationModeUpdateTask to xpack-core
